Update 1.0  - Added Welcome Announcements
This update added a welcome announcement for any member when he/she joins the server, if the server is announcement-enabled (announcement enable command not made yet). If yes, an embed (to be designed, currently placeholder will be sent instead) will be sent  in the announcement channel, which would also have been set when enabling announcements.